Clean up WorkspaceHelper
