+++ org.moflon.ide.plugins/src/org/moflon/util/plugins/PluginProducerWorkspaceRunnable.java
@@ -3,6 +3,7 @@ package org.moflon.util.plugins;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
 import java.util.jar.Manifest;
 
@@ -13,7 +14,13 @@ import org.eclipse.core.resources.IProjectDescription;
 import org.eclipse.core.resources.IWorkspaceRunnable;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IProgressMonitor;
+import org.eclipse.core.runtime.NullProgressMonitor;
+import org.eclipse.core.runtime.Path;
 import org.eclipse.core.runtime.SubMonitor;
+import org.eclipse.jdt.core.IClasspathEntry;
+import org.eclipse.jdt.core.IJavaProject;
+import org.eclipse.jdt.core.JavaCore;
+import org.eclipse.jdt.core.JavaModelException;
 import org.moflon.TGGLanguageActivator;
 import org.moflon.core.moca.tree.MocaTreePlugin;
 import org.moflon.core.utilities.LogUtils;
@@ -61,7 +68,7 @@ public class PluginProducerWorkspaceRunnable implements IWorkspaceRunnable
       try
       {
          addPluginFeatures(project, projectProperties, monitor);
-         WorkspaceHelper.addContainerToBuildPath(project, "org.eclipse.pde.core.requiredPlugins");
+         addContainerToBuildPath(project, "org.eclipse.pde.core.requiredPlugins");
       } catch (IOException e)
       {
          LogUtils.error(logger, e);
@@ -92,8 +99,8 @@ public class PluginProducerWorkspaceRunnable implements IWorkspaceRunnable
          changed |= ManifestFileUpdater.updateAttribute(manifest, PluginManifestConstants.BUNDLE_EXECUTION_ENVIRONMENT,
                properties.get(MetamodelProperties.JAVA_VERION), AttributeUpdatePolicy.KEEP);
 
-         changed |= ManifestFileUpdater.updateDependencies(manifest, Arrays.asList(
-               new String[] { WorkspaceHelper.PLUGIN_ID_ECORE, WorkspaceHelper.PLUGIN_ID_ECORE_XMI, WorkspaceHelper.getPluginId(MoflonUtilitiesActivator.class) }));
+         changed |= ManifestFileUpdater.updateDependencies(manifest, Arrays.asList(new String[] { WorkspaceHelper.PLUGIN_ID_ECORE,
+               WorkspaceHelper.PLUGIN_ID_ECORE_XMI, WorkspaceHelper.getPluginId(MoflonUtilitiesActivator.class) }));
 
          changed |= ManifestFileUpdater.updateDependencies(manifest,
                ManifestFileUpdater.extractDependencies(properties.get(MetamodelProperties.DEPENDENCIES_KEY)));
@@ -180,5 +187,54 @@ public class PluginProducerWorkspaceRunnable implements IWorkspaceRunnable
    private static boolean containsBuilder(final List<ICommand> builders, final String name)
    {
       return builders.stream().anyMatch(c -> c.getBuilderName().equals(name));
+   }
+
+   /**
+    * Adds the given container to the list of build path entries (if not included, yet)
+    */
+   private static void addContainerToBuildPath(final Collection<IClasspathEntry> classpathEntries, final String container)
+   {
+      IClasspathEntry entry = JavaCore.newContainerEntry(new Path(container));
+      for (IClasspathEntry iClasspathEntry : classpathEntries)
+      {
+         if (iClasspathEntry.getPath().equals(entry.getPath()))
+         {
+            // No need to add variable - already on classpath
+            return;
+         }
+      }
+
+      classpathEntries.add(entry);
+   }
+
+   /**
+    * Adds the given container to the build path of the given project if it contains no entry with the same name, yet.
+    */
+   private static void addContainerToBuildPath(final IProject project, final String container)
+   {
+      final IJavaProject iJavaProject = JavaCore.create(project);
+      try
+      {
+         // Get current entries on the classpath
+         Collection<IClasspathEntry> classpathEntries = new ArrayList<>(Arrays.asList(iJavaProject.getRawClasspath()));
+      
+         addContainerToBuildPath(classpathEntries, container);
+      
+         setBuildPath(iJavaProject, classpathEntries);
+      } catch (JavaModelException e)
+      {
+         LogUtils.error(logger, e, "Unable to set classpath variable");
+      }
+   }
+
+   private static void setBuildPath(final IJavaProject javaProject, final Collection<IClasspathEntry> entries) throws JavaModelException
+   {
+      final SubMonitor subMon = SubMonitor.convert(new NullProgressMonitor(), "Set build path", 1);
+      // Create new buildpath
+      IClasspathEntry[] newEntries = new IClasspathEntry[entries.size()];
+      entries.toArray(newEntries);
+      
+      // Set new classpath with added entries
+      javaProject.setRawClasspath(newEntries, subMon.split(1));
    }
 }
